# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "clsschedSectionType",
    "description": "Section Types Data",
    "version": "v1",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.byu.edu",
  "schemes": [
    "https"
  ],
  "basePath": "/domain/clssched/section_types",
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "x-controller": "section_types",
      "get": {
        "summary": "clssched Codes",
        "description": "Retrieve a collection of section_types",
        "operationId": "getSectionTypes",
        "tags": [
          "Section Types"
        ],
        "responses": {
          "200": {
            "description": "Retrieve a collection of section_types",
            "schema": {
              "$ref": "#/definitions/section_types"
            },
            "examples": {
              "application/json": {
                "$ref": "#/x-mock_json/section_types"
              }
            }
          },
          "400": {
            "$ref": "#/responses/400"
          },
          "401": {
            "$ref": "#/responses/401"
          },
          "403": {
            "$ref": "#/responses/403"
          },
          "409": {
            "$ref": "#/responses/409"
          },
          "default": {
            "$ref": "#/responses/default"
          }
        }
      },
      "post" : {
        "summary": "clssched Codes",
        "description": "Creates a new section type",
        "operationId": "createSectionType",
        "parameters": [
          {
            "name": "section_type_new",
            "in": "body",
            "description": "New section type values",
            "required": true,
            "schema": {
              "$ref": "#/definitions/section_type_new"
            }
          }
        ],
        "tags": [
          "Section Types"
        ],
        "responses": {
          "201": {
            "description": "Section Type successfully created",
            "schema": {
              "$ref": "#/definitions/section_type_data"
            },
            "examples": {
              "application/json": {
                "$ref": "#/x-mock_json/section_types/values/0"
              }
            }
          },
          "400": {
            "$ref": "#/responses/400"
          },
          "401": {
            "$ref": "#/responses/401"
          },
          "403": {
            "$ref": "#/responses/403"
          },
          "409": {
            "$ref": "#/responses/409"
          },
          "default": {
            "$ref": "#/responses/default"
          }
        }
      }
    },
    "/{section_type}": {
      "x-controller": "section_type",
      "get": {
        "summary": "clssched Codes",
        "description": "Retrieve a specific section_type",
        "operationId": "getSectionType",
        "parameters": [
          {
            "$ref": "#/parameters/section_type"
          }
        ],
        "tags": [
          "Section Types"
        ],
        "responses": {
          "200": {
            "description": "A specific section_type",
            "schema": {
              "$ref": "#/definitions/section_type_data"
            },
            "examples": {
              "application/json": {
                "$ref": "#/x-mock_json/section_types/values/0"
              }
            }
          },
          "400": {
            "$ref": "#/responses/400"
          },
          "401": {
            "$ref": "#/responses/401"
          },
          "403": {
            "$ref": "#/responses/403"
          },
          "404": {
            "description": "There is no matching section_type",
            "schema": {
              "$ref": "#/definitions/error_response"
            }
          },
          "409": {
            "$ref": "#/responses/409"
          },
          "default": {
            "$ref": "#/responses/default"
          }
        }
      },
      "put": {
        "summary": "clssched Codes",
        "description": "Modify a section_type",
        "operationId": "modifySectionType",
        "parameters": [
          {
            "$ref": "#/parameters/section_type"
          },
          {
            "name": "section_type_modify",
            "in": "body",
            "description": "Values required to modify a section_type",
            "required": true,
            "schema": {
              "$ref": "#/definitions/section_type_new"
            }
          }
        ],
        "tags": [
          "Section Types"
        ],
        "responses": {
          "200": {
            "description": "section_type successfully Modified",
            "schema": {
              "$ref": "#/definitions/section_type_data"
            },
            "examples": {
              "application/json": {
                "$ref": "#/x-mock_json/section_types/values/0"
              }
            }
          },
          "400": {
            "$ref": "#/responses/400"
          },
          "401": {
            "$ref": "#/responses/401"
          },
          "403": {
            "$ref": "#/responses/403"
          },
          "404": {
            "description": "There is no matching section_type",
            "schema": {
              "$ref": "#/definitions/error_response"
            }
          },
          "409": {
            "$ref": "#/responses/409"
          },
          "default": {
            "$ref": "#/responses/default"
          }
        }
      },
      "delete": {
        "summary": "clssched Codes",
        "description": "Removes a section_type",
        "operationId": "removeSectionType",
        "parameters": [
          {
            "$ref": "#/parameters/section_type"
          }
        ],
        "tags": [
          "Section Types"
        ],
        "responses": {
          "204": {
            "description": "section_type successfully Removed"
          },
          "400": {
            "$ref": "#/responses/400"
          },
          "401": {
            "$ref": "#/responses/401"
          },
          "403": {
            "$ref": "#/responses/403"
          },
          "default": {
            "$ref": "#/responses/default"
          }
        }
      }
    }
  },
  "definitions": {
    "section_types": {
      "description": "A collection of section types",
      "type": "object",
      "required": [
        "links",
        "metadata",
        "values"
      ],
      "properties": {
        "links": {
          "description": "Hypermedia Links that impact the section type collection",
          "type": "object",
          "required": [
            "section_types__info"
          ],
          "properties": {
            "section_types__info": {
              "type": "object",
              "required": [
                "rel",
                "href",
                "method"
              ],
              "properties": {
                "rel": {
                  "type": "string",
                  "enum": [
                    "self"
                  ],
                  "default": "self"
                },
                "href": {
                  "type": "string",
                  "format": "uri",
                  "default": "https://api.byu.edu/domain/clssched/section_types/v1"
                },
                "method": {
                  "type": "string",
                  "enum": [
                    "GET"
                  ],
                  "default": "GET"
                }
              },
              "example": {
                "$ref": "#/x-mock_json/section_types/links/section_types__info"
              }
            },
            "section_types__create": {
              "type": "object",
              "required": [
                "rel",
                "href",
                "method"
              ],
              "properties": {
                "rel": {
                  "type": "string",
                  "enum": [
                    "section_types__create"
                  ],
                  "default": "section_types__create"
                },
                "href": {
                  "type": "string",
                  "format": "uri",
                  "default": "https://api.byu.edu/domain/clssched/section_types/v1/"
                },
                "method": {
                  "type": "string",
                  "enum": [
                    "POST"
                  ],
                  "default": "POST"
                }
                },
              "example": {
                "$ref": "#/x-mock_json/section_types/links/section_types__create"
              }
            }
          },
          "example": {
            "$ref": "#/x-mock_json/section_types/links"
          }
        },
        "metadata": {
          "$ref": "#/definitions/collection_metadata"
        },
        "values": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/section_type_data"
          },
          "x-variable": "values"
        }
      },
      "example": {
        "$ref": "#/x-mock_json/section_types"
      }
    },
    "section_type_data": {
      "description": "A section type",
      "type": "object",
      "properties": {
        "links": {
          "description": "Hypermedia Links that impact this section type",
          "type": "object",
          "required": [
            "section_types__info"
          ],
          "properties": {
            "section_type__info": {
              "description": "Hypermedia link to retrieve this section_type",
              "type": "object",
              "required": [
                "rel",
                "href",
                "method"
              ],
              "properties": {
                "rel": {
                  "type": "string",
                  "enum": [
                    "self"
                  ],
                  "default": "self"
                },
                "href": {
                  "type": "string",
                  "format": "uri",
                  "default": "https://api.byu.edu/domain/clssched/section_types/v1/{section_type}",
                  "x-template": "https://api.byu.edu/domain/clssched/section_types/v1/:section_type"
                },
                "method": {
                  "type": "string",
                  "enum": [
                    "GET"
                  ],
                  "default": "GET"
                }
              },
              "example": {
                "$ref": "#/x-mock_json/section_types/values/0/links/section_type__info"
              }
            },
            "section_type__modify": {
              "description": "Hypermedia link to modify this section_type",
              "type": "object",
              "required": [
                "rel",
                "href",
                "method"
              ],
              "properties": {
                "rel": {
                  "type": "string",
                  "enum": [
                    "section_type__modify"
                  ],
                  "default": "section_type__modify"
                },
                "href": {
                  "type": "string",
                  "format": "uri",
                  "default": "https://api.byu.edu/domain/clssched/section_types/v1/{section_type}",
                  "x-template": "https://api.byu.edu/domain/clssched/section_types/v1/:section_type"
                },
                "method": {
                  "type": "string",
                  "enum": [
                    "PUT"
                  ],
                  "default": "PUT"
                }
              },
              "example": {
                "$ref": "#/x-mock_json/section_types/values/0/links/section_type__modify"
              }
            },
            "section_type__remove": {
              "description": "Hypermedia link to remove this section_type",
              "type": "object",
              "required": [
                "rel",
                "href",
                "method"
              ],
              "properties": {
                "rel": {
                  "type": "string",
                  "enum": [
                    "section_type__remove"
                  ],
                  "default": "section_type__remove"
                },
                "href": {
                  "type": "string",
                  "format": "uri",
                  "default": "https://api.byu.edu/domain/clssched/section_types/v1/{section_type}",
                  "x-template": "https://api/byu.edu/domain/clssched/section_types/v1/:section_type"
                },
                "method": {
                  "type": "string",
                  "enum": [
                    "DELETE"
                  ],
                  "default": "DELETE"
                }
              },
              "example": {
                "$ref": "#/x-mock_json/section_types/values/0/links/section_type__remove"
              }
            }
          },
          "example": {
            "$ref": "#/x-mock_json/section_types/values/0/links"
          }
        },
        "metadata": {
          "$ref": "#/definitions/simple_metadata"
        },
        "section_type": {
          "$ref": "#/definitions/section_type"
        },
        "description": {
          "$ref": "#/definitions/description"
        },
        "long_description": {
          "$ref": "#/definitions/long_description"
        },
        "active_status": {
          "$ref": "#/definitions/active_status"
        },
        "updated_by_id": {
          "$ref": "#/definitions/updated_by_id"
        },
        "date_time_updated": {
          "$ref": "#/definitions/date_time_updated"
        }
      },
      "example": {
        "$ref": "#/x-mock_json/section_types/values/0"
      }
    },
    "section_type_new": {
      "description": "Elements required to add a new section_type",
      "type": "object",
      "required": [
        "section_type",
        "description",
        "long_description",
        "active_status"
      ],
      "properties": {
        "section_type": {
          "$ref": "#/definitions/section_type"
        },
        "description": {
          "$ref": "#/definitions/description"
        },
        "long_description": {
          "$ref": "#/definitions/long_description"
        },
        "active_status": {
          "$ref": "#/definitions/active_status"
        }
      },
      "example": {
        "$ref": "#/x-mock_json/section_type_new"
      }
    },
    "error_response": {
      "type": "object",
      "required": [
        "metadata"
      ],
      "properties": {
        "metadata": {
          "$ref": "#/definitions/simple_metadata"
        }
      }
    },
    "simple_metadata": {
      "type": "object",
      "required": [
        "validation_response"
      ],
      "properties": {
        "validation_response": {
          "type": "object",
          "required": [
            "code",
            "message"
          ],
          "properties": {
            "code": {
              "type": "integer",
              "default": 200,
              "x-variable": "code",
              "example": "#/x-mock_json/section_types/values/0/metadata/validation_response/code"
            },
            "message": {
              "type": "string",
              "default": "Successful",
              "x-variable": "message",
              "example": "#/x-mock_json/section_types/values/0/metadata/validation_response/message"
            }
          },
          "example": {
            "$ref": "#/x-mock_json/section_types/values/0/metadata/validation_response"
          }
        },
        "validation_information": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "cache": {
          "description": "If data is from live source, value will be null, otherwise it will be date-time of cache",
          "type": "object",
          "properties": {
            "date_time": {
              "type": "string",
              "format": "date-time"
            }
          }
        }
      },
      "example": {
        "$ref": "#/x-mock_json/section_types/values/0/metadata"
      }
    },
    "collection_metadata": {
      "allOf": [
        {
          "$ref": "#/definitions/simple_metadata"
        },
        {
          "type": "object",
          "required": [
            "collection_size"
          ],
          "properties": {
            "collection_size": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "x-variable": "collection_size",
              "example": {
                "$ref": "#/x-mock_json/section_types/metadata/collection_size"
              }
            }
          }
        }
      ],
      "example": {
        "$ref": "#/x-mock_json/section_types/metadata"
      }
    },
    "section_type": {
      "description": "section_type",
      "type": "string",
      "example": {
        "$ref": "#/x-mock_json/section_types/values/0/section_type"
      },
      "x-variable": "section_type"
    },
    "active_status": {
      "description": "Is the code currently active or only used for archive purposes?",
      "type": "string",
      "enum": [
        "Y",
        "N"
      ],
      "example": {
        "$ref": "#/x-mock_json/section_types/values/0/active_status"
      },
      "x-variable": "active_status"
    },
    "date_time_updated": {
      "description": "When was the section_type's data last updated?",
      "type": "string",
      "format": "date-time",
      "example": {
        "$ref": "#/x-mock_json/section_types/values/0/date_time_updated"
      },
      "x-variable": "date_time_updated"
    },
    "description": {
      "description": "section_type short description",
      "type": "string",
      "example": {
        "$ref": "#/x-mock_json/section_types/values/0/description"
      },
      "x-variable": "description"
    },
    "long_description": {
      "description": "section_type long description",
      "type": "string",
      "example": {
        "$ref": "#/x-mock_json/section_types/values/0/long_description"
      },
      "x-variable": "long_description"
    },
    "updated_by_id": {
      "description": "Who last updated the section_type's data?",
      "type": "string",
      "example": {
        "$ref": "#/x-mock_json/section_types/values/0/updated_by_id"
      },
      "x-variable": "updated_by_id"
    }
  },
  "parameters": {
    "section_type": {
      "name": "section_type",
      "in": "path",
      "description": "section_type path parameter",
      "required": true,
      "type": "string"
    }
  },
  "responses": {
    "400": {
      "description": "Invalid Request - unable to interpret request",
      "schema": {
        "$ref": "#/definitions/error_response"
      }
    },
    "401": {
      "description": "Authentication required to access data",
      "schema": {
        "$ref": "#/definitions/error_response"
      }
    },
    "403": {
      "description": "Access denied to protected data",
      "schema": {
        "$ref": "#/definitions/error_response"
      }
    },
    "409": {
      "description": "Query parameters conflict or are malformed",
      "schema": {
        "$ref": "#/definitions/error_response"
      }
    },
    "default": {
      "description": "Unexpected error",
      "schema": {
        "$ref": "#/definitions/error_response"
      }
    }
  },
  "x-mock_json": {
    "section_types": {
      "links": {
        "section_types__info": {
          "rel": "self",
          "href": "https://api.byu.edu/domain/clssched/section_types/v1",
          "method": "GET"
        },
        "section_types__create": {
          "rel": "section_types__create",
          "href": "https://api.byu.edu/domain/clssched/section_types/v1/",
          "method": "POST"
        }
      },
      "metadata": {
        "collection_size": 1,
        "validation_response": {
          "code": 200,
          "message": "Retrieve Successful"
        }
      },
      "values": [
        {
          "links": {
            "section_type__info": {
              "rel": "self",
              "href": "https://api.byu.edu/domain/clssched/section_types/v1/IDAHO",
              "method": "GET"
            },
            "section_type__modify": {
              "rel": "section_type__modify",
              "href": "https://api.byu.edu/domain/clssched/section_types/v1/IDAHO",
              "method": "PUT"
            },
            "section_type__remove": {
              "rel": "section_type__remove",
              "href": "https://api.byu.edu/domain/clssched/section_types/v1/IDAHO",
              "method": "DELETE"
            }
          },
          "metadata": {
            "validation_response": {
              "code": 200,
              "message": "Successful"
            }
          },
          "section_type": "IDAHO",
          "description": "Idaho Center",
          "long_description": "Idaho Center",
          "active_status": "Y",
          "updated_by_id": "INIT",
          "date_time_updated": "2005-11-01T00:00:00.000Z"
        }
      ]
    },
    "section_type_new": {
      "section_type": "IDAHO",
      "description": "Idaho Center",
      "long_description": "Idaho Center",
      "active_status": "Y"
    }
  }
}